requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
